Reorganize legacy charmrepo code.
